server/requirements.txt
flask
flask-socketio
requests
gevent
gevent-websocket
gunicorn
Flask-PyMongo
ansicolors
rfc3339
